update quantification summary workflow content (#182)
* update quantification challenge directory structure

* update summary workflow content

* address suggestions from CJ

* consistent --challenge_type description with compute_metrics.py suggested by Asier

* address suggestions from Asier

* address suggestions from Asier

* address suggestions from Asier

* DEFAULT_eventMark -> event_date and docker container directories (apaeval_* -> q2_*)

* update build.sh and hopefully solve merge conflict

* Clarify that by default parameters are read from nextflow.config

* Clarify that in default mode parameters are read from nextflow.config

* address comments from Weronika, Dominik, and Asier

* remove Q2_specification/

* extend window param to fully work with workflow

* Cj summary wf cleanup (#215)

* clean validation

* adressed comment from #144

* clean metrics computation

* clean consolidation

* READMEs and renaming

* main summary wf README

* add TOC

* update quantification README

* update quantification_dockers README

* add OEB DB benchmarking data query/download

* formatting

* docs: assert presence of input files to avoid totally meaningless error msg

* docs: comments in code

* refactor: rename variables for better understanding

* refactor: parameterize title and axis labels

* refactor: rename Q specification

* refactor: convenient local usage
