+++ summary_workflows/quantification/nextflow.config
@@ -12,17 +12,17 @@ profiles {
    docker {
       process {
           withName: validation{
-            container = "docker.io/apaeval/q2_validation:1.0"
+            container = "docker.io/apaeval/q_validation:1.0"
           }
       }
       process {
           withName: compute_metrics{
-            container = "docker.io/apaeval/q2_metrics:1.0"
+            container = "docker.io/apaeval/q_metrics:1.0"
           }
       }
       process {
           withName: benchmark_consolidation{
-            container = "docker.io/apaeval/q2_consolidation:1.0"
+            container = "docker.io/apaeval/q_consolidation:1.0"
           }
       }
       
@@ -37,22 +37,19 @@ profiles {
 params  {
 
   // submitted file
-  input = "$baseDir/APAeval_Q2_test_data/embryonicCortex.PAPERCLIP.mm10.bed"
-
-  // reference file dir to validate input data
-  public_ref_dir = "$baseDir/APAeval_Q2_test_data/public_ref"
+  input = "$baseDir/APAeval_Q_test_data/embryonicCortex.PAPERCLIP.mm10.bed"
 
   // name of the tool used for the predicitions
-  participant_id = "test"
+  participant_id = "tool1"
 
   // directory where the 'gold standards' are found
-  goldstandard_dir = "$baseDir/APAeval_Q2_test_data/metrics_ref_datasets"
+  goldstandard_dir = "$baseDir/APAeval_Q_test_data/metrics_ref_datasets"
 
-  // challenge types for which the benchmark has to be performed
-  challenges_ids  = "Q2"
+  // Challenge IDs (separated by spaces) for which the benchmark has to be performed
+  challenge_ids  = "Qx"
 
   // directory where APAeval benchmarking data is found
-  assess_dir = "$baseDir/APAeval_Q2_test_data/data"
+  assess_dir = "$baseDir/APAeval_Q_test_data/data"
 
   //name or OEB permanent ID for the benchmarking community
   community_id = "APAeval"
@@ -61,14 +58,20 @@ params  {
   // challenge_status = "CLOSED"
 
   // directories where results will be written
-  validation_result = "sample_out/participant_out/validated-participant-data.json"
-  assessment_results = "sample_out/assessment_out/Assessment_datasets.json"
-  outdir = "sample_out/results"
-  statsdir = "sample_out/stats"
-  outdir = "sample_out/results"
-  data_model_export_dir = "sample_out/benchmarking_data_model_export/consolidated_result.json"
-  otherdir = "sample_out/other"
+  output = "${participant_id}_benchmark"
+  validation_result = "${params.output}/validated_participant_data.json"
+  assessment_results = "${params.output}/assessment_datasets.json"
+  outdir = "${params.output}/results"
+  statsdir = "${params.output}/stats"
+  consolidation_result = "${params.output}/consolidated_result.json"
+  otherdir = "${params.output}/other"
 
+  // other parameters
+  window = 15
+  offline = 1 // type int; 0 for False
+
+  //set DEFAULT_eventMark
+  event_date = '2021-06-04'
 }
 
 // By default output execution reports
