+++ summary_workflows/quantification/quantification_dockers/build.sh
@@ -10,8 +10,8 @@ fi
 if [ $# -gt 0 ]; then
 	tag_id="$1"
 
-	for docker_name in tcga_validation tcga_metrics tcga_consolidation ; do
-		docker build -t "$docker_name":"$tag_id" "$docker_name"
+	for docker_name in q_validation q_metrics q_consolidation ; do
+		docker build -t apaeval/"$docker_name":"$tag_id" "$docker_name"
 	done
 else
 	echo "Usage: $0 tag_id" 1>&2
