+++ summary_workflows/quantification/quantification_dockers/q_consolidation/merge_data_model_files.py
@@ -9,9 +9,10 @@ from argparse import ArgumentParser
 
 def main(args):
     # input parameters
-    metrics_dir = args.metrics_data
-    participant_dir = args.participant_data
-    aggregation_dir = args.aggregation_data
+    metrics_data = args.metrics_data
+    validation_data = args.validation_data
+    aggregation_data = args.aggregation_data
+    challenge = args.challenge_id
     out_path = args.output
 
     # Assuring the output path does exist
@@ -22,12 +23,17 @@ def main(args):
                 pass
         except OSError as exc:
             print("OS error: {0}".format(exc) + "\nCould not create output path: " + out_path)
-
+            
+    # This is the final consolidated output
     data_model_file = []
 
-    data_model_file = join_json_files(participant_dir, data_model_file, "*.json")
-    data_model_file = join_json_files(metrics_dir, data_model_file, "*.json")
-    data_model_file = join_json_files(aggregation_dir, data_model_file, "*_summary.json")
+    # get the output files from previous steps and concatenate
+    # from validation ("validated_participant_data")
+    data_model_file = join_json_files(validation_data, data_model_file, "*.json")
+    # from metrics ("assessment_out")
+    data_model_file = join_json_files(metrics_data, data_model_file, "*.json")
+    # from consolidation part 1 (manage_assessment_data.py), "sample_out/results/challenge.json"
+    data_model_file = join_json_files(aggregation_data, data_model_file, "*" + challenge + ".json")
 
     # write the merged data model file to json output
     with open(out_path, mode='w', encoding="utf-8") as f:
@@ -65,10 +71,11 @@ def join_json_files(data_directory, data_model_file, file_extension):
 if __name__ == '__main__':
 
     parser = ArgumentParser()
-    parser.add_argument("-p", "--participant_data", help="path where the data for the participant is stored", required=True)
-    parser.add_argument("-m", "--metrics_data", help="path where the data for the assessment metrics are stored", required=True)
+    parser.add_argument("-v", "--validation_data", help="path to validated_participant_data.json", required=True)
+    parser.add_argument("-m", "--metrics_data", help="path to assessment_datasets.json", required=True)
     parser.add_argument("-a", "--aggregation_data", help="dir where the data for benchmark summary/aggregation are stored",
                         required=True)
+    parser.add_argument("-c", "--challenge_id", help="Id of the challenge", required=True)
     parser.add_argument("-o", "--output", help="output path where the minimal dataset JSON file will be written", required=True)
 
     args = parser.parse_args()
