+++ summary_workflows/quantification/quantification_dockers/q_consolidation/requirements.txt
@@ -1,3 +1,4 @@
 matplotlib
 numpy
 pandas
+requests
