+++ summary_workflows/quantification/quantification_dockers/q_validation/.dockerignore
