+++ summary_workflows/quantification/quantification_dockers/q_validation/JSON_templates/JSON_templates.py
