+++ summary_workflows/quantification/specification/example_files/input2.bed
